pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "daily-podcast"
version = "0.1.0"
description = "Daily arXiv-to-podcast pipeline using Gmail, arXiv PDFs, and NotebookLM automation."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.32.0",
  "python-dotenv>=1.0.0",
  "google-api-python-client>=2.170.0",
  "google-auth-httplib2>=0.2.0",
  "google-auth-oauthlib>=1.2.0",
  "playwright>=1.52.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
daily_podcast = "daily_podcast.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
